+++ pkgs/watcher/lib/src/directory_watcher/macos/native_watch.dart
@@ -6,15 +6,19 @@ import 'dart:async';
 import 'dart:io';
 
 import '../../event.dart';
-import '../../unix_paths.dart';
-import '../../utils.dart';
-import 'event_tree.dart';
+import '../../event_batching.dart';
+import '../../paths.dart';
+import '../../testing.dart';
+import '../event_tree.dart';
 
-/// Watches a directory tree with the native MacOS watcher.
+/// Recursively watches a directory with `Directory.watch` on MacOS or Windows.
 ///
 /// Handles incorrect closure of the watch due to a delete event from before
 /// the watch started, by re-opening the watch if the directory still exists.
 /// See https://github.com/dart-lang/sdk/issues/14373.
+///
+/// Handles deletion of the watched directory on Windows by watching the parent
+/// directory.
 class NativeWatch {
   final AbsolutePath watchedDirectory;
 
@@ -30,7 +34,11 @@ class NativeWatch {
   /// Called with native watch errors.
   final void Function(Object, StackTrace) _onError;
 
+  /// The underlying batched event stream.
   StreamSubscription<List<Event>>? _subscription;
+
+  /// On Windows only, the parent directory event stream.
+  StreamSubscription<FileSystemEvent>? _parentSubscription;
 
   /// Watches [watchedDirectory].
   ///
@@ -47,14 +55,64 @@ class NativeWatch {
         _watchedDirectoryWasDeleted = watchedDirectoryWasDeleted {
     logForTesting?.call('NativeWatch(),$watchedDirectory');
     _watch();
+    if (Platform.isWindows) _watchParent();
   }
 
   void _watch() {
     _subscription?.cancel();
-    _subscription = watchedDirectory
-        .watch(recursive: true)
-        .batchAndConvertEvents()
-        .listen(_onData, onError: _onError, onDone: _onDone);
+    // In older SDKs watcher exceptions on Windows are not sent over the stream
+    // and must be caught with a zone handler.
+    runZonedGuarded(
+      () {
+        _subscription = watchedDirectory
+            .watch(recursive: true)
+            .batchAndConvertEventsForPlatform()
+            .listen(_onData,
+                onError: _restartWatchOnOverflowOr(_onError), onDone: _onDone);
+      },
+      _restartWatchOnOverflowOr(Error.throwWithStackTrace),
+    );
+  }
+
+  /// Handles deletes and moves of [watchedDirectory] on Windows.
+  ///
+  /// Deletes can be signalled by an exception, but moves are not signalled
+  /// at all. So, handle both by watching the parent directory.
+  ///
+  /// See https://github.com/dart-lang/sdk/issues/62193 and
+  /// https://github.com/dart-lang/sdk/issues/62194.
+  void _watchParent() {
+    final parent = watchedDirectory.parent;
+    if (parent == watchedDirectory) {
+      // Watching a filesystem root: it can't be deleted.
+      return;
+    }
+    final parentStream = parent.watch(recursive: false);
+    _parentSubscription = parentStream.listen(
+      (event) {
+        // Only look at events for [watchedDirectory].
+        final eventPath = AbsolutePath(event.path);
+        if (eventPath.basename != watchedDirectory.basename) {
+          return;
+        }
+        // The directory was deleted if there is an event saying it was deleted,
+        // or if there was any event and it no longer exists. Note that it might
+        // still exist but be a different+new directory: this needs handling as
+        // a delete because the new directory would need a new native watch.
+        if (event is FileSystemMoveEvent ||
+            event is FileSystemDeleteEvent ||
+            (eventPath.typeSync() == FileSystemEntityType.notFound)) {
+          _watchedDirectoryWasDeleted();
+        }
+      },
+      onError: (error) {
+        // Ignore errors, simply close the stream. The user listens on
+        // [directory], and while it can fail to listen on the parent, we may
+        // still be able to listen on the path requested.
+        _parentSubscription?.cancel();
+        _parentSubscription = null;
+      },
+    );
   }
 
   /// Closes the watch.
@@ -62,6 +120,8 @@ class NativeWatch {
     logForTesting?.call('NativeWatch,$watchedDirectory,close');
     _subscription?.cancel();
     _subscription = null;
+    _parentSubscription?.cancel();
+    _parentSubscription = null;
   }
 
   void _onData(List<Event> events) {
@@ -73,7 +133,15 @@ class NativeWatch {
           event.absolutePath == watchedDirectory) {
         continue;
       }
-      eventTree.add(event.absolutePath.relativeTo(watchedDirectory));
+      // Drop paths outside the watched directory, which could only be due to
+      // a move event destination path. Currently the VM reports moves to
+      // outside the watched directory as deletes, so there aren't any such move
+      // events, but it's as easy and more future proof to handle correctly by
+      // dropping instead of throwing.
+      final path = event.absolutePath.tryRelativeTo(watchedDirectory);
+      if (path != null) {
+        eventTree.add(path);
+      }
     }
     _onEvents(eventTree);
   }
@@ -88,5 +156,61 @@ class NativeWatch {
     } else {
       _watchedDirectoryWasDeleted();
     }
+  }
+
+  /// Intercepts and handles Windows-specific exceptions.
+  ///
+  /// A "closed unexpectedly" error happens on Windows when the event
+  /// stream is not serviced quickly enough and the OS buffer fills.
+  ///
+  /// And, a `SocketException` happens on Windows when the watched directory
+  /// is deleted.
+  void Function(Object, StackTrace) _restartWatchOnOverflowOr(
+      void Function(Object, StackTrace) otherwise) {
+    return (error, stackTrace) async {
+      if (error is FileSystemException &&
+          error.message.startsWith('Directory watcher closed unexpectedly')) {
+        // Wait to work around https://github.com/dart-lang/sdk/issues/61378.
+        // Give the VM time to reset state after the error. See the issue for
+        // more discussion of the workaround.
+        // TODO(davidmorgan): remove the wait once min SDK version is 3.10.
+        // The recovery test in `windows_isolate_test.dart` will continue to
+        // pass if it's no longer needed.
+        await _subscription?.cancel();
+        await Future<void>.delayed(const Duration(milliseconds: 1));
+        _watch();
+        _watchedDirectoryWasRecreated();
+      } else if ((error is SocketException &&
+              error.message.contains('SocketException: Access is denied')) ||
+          (error is FileSystemException &&
+              error.message.contains('SocketException: Access is denied'))) {
+        // This can happen if the watched directory is deleted, see
+        // [_watchParent] which handles both deletes and moves. Ignore the
+        // exception.
+      } else {
+        otherwise(error, stackTrace);
+      }
+    };
+  }
+}
+
+extension _BatchEvents on Stream<FileSystemEvent> {
+  /// Batches events based on the current platform.
+  ///
+  /// On Windows, events need to be batched by path for two reasons: to handle
+  /// duplicate events together and because polling the filesystem state too
+  /// quickly after an event arrives can give incorrect results. In particular,
+  /// if the entity is a newly-created link to a directory then polling too soon
+  /// reports that it is a directory, not a link. By testing, a 1ms delay looks
+  /// sufficient: incorrect type dropped from 150/1000 to 0/10000. Use a 5ms
+  /// delay to have a margin for error for load and machine performance.
+  ///
+  /// On other platforms, which means MacOS, events are batched by "nearby
+  /// microtask" to pick up all the events that the VM sends "together".
+  Stream<List<Event>> batchAndConvertEventsForPlatform() {
+    return Platform.isWindows
+        ? batchBufferedByPathAndConvertEvents(
+            duration: const Duration(milliseconds: 5))
+        : batchNearbyMicrotasksAndConvertEvents();
   }
 }