+++ pkgs/watcher/lib/src/event.dart
@@ -40,6 +40,21 @@ extension type Event._(FileSystemEvent _event) {
     return result;
   }
 
+  /// Returns an iterable containing this event, split to a "create" and a
+  /// "delete" event if it's a move event.
+  Iterable<Event> splitIfMove() sync* {
+    if (type != EventType.moveFile && type != EventType.moveDirectory) {
+      yield this;
+      return;
+    }
+    final destination = this.destination;
+    yield Event._(FileSystemDeleteEvent(path, type == EventType.moveDirectory));
+    if (destination != null) {
+      yield Event._(
+          FileSystemCreateEvent(destination, type == EventType.moveDirectory));
+    }
+  }
+
   /// A create event for a file at [path].
   static Event createFile(String path) =>
       Event._(FileSystemCreateEvent(path, false));