+++ pkgs/watcher/test/directory_watcher/event_tree_test.dart
@@ -2,9 +2,13 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:io';
+
 import 'package:test/test.dart';
-import 'package:watcher/src/directory_watcher/macos/event_tree.dart';
-import 'package:watcher/src/unix_paths.dart';
+import 'package:watcher/src/directory_watcher/event_tree.dart';
+import 'package:watcher/src/paths.dart';
+
+final separator = Platform.pathSeparator;
 
 void main() {
   group('EventTree', () {
@@ -28,7 +32,7 @@ void main() {
 
     test('event tree with event deep under root has expected single event', () {
       final eventTree = EventTree();
-      eventTree.add(RelativePath('a/b'));
+      eventTree.add(RelativePath('a${separator}b'));
       expect(eventTree.isSingleEvent, false);
 
       expect(eventTree[PathSegment('a')]!.isSingleEvent, false);
@@ -38,7 +42,7 @@ void main() {
 
     test('adding event removes tree under it', () {
       final eventTree = EventTree();
-      eventTree.add(RelativePath('a/b'));
+      eventTree.add(RelativePath('a${separator}b'));
       eventTree.add(RelativePath('a'));
 
       expect(eventTree[PathSegment('a')]![PathSegment('b')], null);
@@ -47,7 +51,7 @@ void main() {
     test("events can't be added under an event", () {
       final eventTree = EventTree();
       eventTree.add(RelativePath('a'));
-      eventTree.add(RelativePath('a/b'));
+      eventTree.add(RelativePath('a${separator}b'));
 
       expect(eventTree[PathSegment('a')]![PathSegment('b')], null);
     });